kryotecsense/server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.23.5
alembic==1.13.1
bcrypt==3.2.0
aio-pika==9.4.0
# Dependencias para generación de reportes
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.4
